libcamera: Cleanup color format organization
Driver only supports a single color format, simplify the code to reflect.
Driver abstraction should only be concerned with driver format, not
application or HAL's output formats.

ControlThread is updated to control the color format selection for all
subcomponents. This makes it easier for the ControlThread to set up
the data flow path and route data.

Change-Id: I071257fb6be6e393531879bebf4d0cfc28e5e348
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>