libcamera: Add PipeThread
Add PipeThread to pipe data from the ControlThread to the PreviewThread
and VideoThread. The PipeThread is responsible for color conversions
from the driver output format to the preview and video input formats.

The PipeThread was introduced to so that a single color conversion
could be done since the preview callback and video use the same
color format. In this implementation, preview and video share the
same buffer.

The buffer that is used for display is still kept internal to the
PreviewThread.

Change-Id: I72861eaf3c5913840660d02392ea9a9d2a7df413
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>