libcamera: HAL callback rework
Rework the logic for the way that buffers are returned to the ControlThread.
Use a single callback in the ControlThread and specify buffer types so
ControlThread can make the appropriate action. This patch also adds a second,
optional buffer parameter to the buffer callback and adds some whitespace
to make code more legible.

Change-Id: I48977b623a5f221fa3a666627f1c758c1ce384f4
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>