CameraBuffer: refactory
CameraBuffer is now a class.

They are passed by reference/pointer in general to
different processing threads. The ControlThread and
CameraDriver work together to  keep all those buffer
objects in a pool and won't destroy them until
a session ends

Methods are created for memory allocation
and user pointer retrieval, which will be needed for
buffer sharing.

Change-Id: I58cc60673b212b7369221a65ba9a230f9c2af3e1