CameraBuffer: implement reader reference count
A simple reference count scheme to allow readers to hold
buffer for processing. Readers decrement the count when done
with the buffer. When reference goes to zero, the buffer is
automatically queued back into driver for writing.

Removed coupled buffer because it no longer is needed.
Readers also do not determine when to return buffer. They only
need to increment/decrement.

Change-Id: I237410f213fddc9269be386add4a9d9fe6f7e5a1