Workaround camera HAL for better snapshot results
Because of some camera sensors which give bad frames after imediate
start of the driver, we added a workaround in the camera HAL so that when
taking pictures we don't stop the driver and use the current frame as
the picture frame.
Also, when taking pictures, because camera application calls stopPreview
then startPreview, we fake the stop and start preview in order to
avoid stoping and starting of camera driver (therefore avoiding bad
frames).
This patch needs to be reworked in order to handle untested situations
(for example if a camera app does not call stopPreview followed by a
startPreview after takePicture).

Change-Id: I750c4d935b7640f4249ab18aab3ac89fc1057417
Signed-off-by: Robert Chiras <robert.chiras@intel.com>
Reviewed-on: https://android.intel.com/176555
Reviewed-by: Serban, Mihai <mihai.serban@intel.com>
Tested-by: Serban, Mihai <mihai.serban@intel.com>